+++ srcs/ft_printf/utils/int_wstrlen.c
@@ -6,7 +6,7 @@
 /*   By: jodufour <jodufour@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/05 15:39:10 by jodufour          #+#    #+#             */
-/*   Updated: 2021/11/15 23:59:34 by jodufour         ###   ########.fr       */
+/*   Updated: 2021/11/20 07:21:55 by jodufour         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,5 +18,5 @@ int	int_wstrlen(wchar_t const *str)
 
 	while (*ptr)
 		++ptr;
-	return (ptr - str);
+	return ((int)(ptr - str));
 }
